% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Background_sampling}
\alias{Background_sampling}
\title{sampling cells for reference data construction
input- cell_clusterid_mat: matrix with columns --"cell.name(UMI)"--same as colnames of counts matrix,"cluster.id","Tissue"; counts_norm: normalised umi counts matrix-- genes X UMIs; ncells.sample: number of cells to sample; clusterids_vec: vector with elements corresponding column names in cell_clusterid_mat with the correct heirarchiel order of clusterids (eg. Tissue, cell type, cell subtype)
output- list: for every gene, sampled gene expression vector}
\usage{
Background_sampling(
  counts_norm_mat,
  cell_clusterid_mat,
  ncells.sample = 50,
  clusterids_vec = c("Tissue", "clusterid.1", "clusterid.2"),
  min.cells.gexp = 10
)
}
\description{
sampling cells for reference data construction
input- cell_clusterid_mat: matrix with columns --"cell.name(UMI)"--same as colnames of counts matrix,"cluster.id","Tissue"; counts_norm: normalised umi counts matrix-- genes X UMIs; ncells.sample: number of cells to sample; clusterids_vec: vector with elements corresponding column names in cell_clusterid_mat with the correct heirarchiel order of clusterids (eg. Tissue, cell type, cell subtype)
output- list: for every gene, sampled gene expression vector
}
