% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{classify}
\alias{classify}
\title{Classify gene activity
input- output of compute_pval, thresholds
output- list: discretised data matrix: genes x cells, significance values}
\usage{
classify(sig_vals, th, th_i)
}
\description{
Classify gene activity
input- output of compute_pval, thresholds
output- list: discretised data matrix: genes x cells, significance values
}
