% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{combine.subset}
\alias{combine.subset}
\title{Combine data lists - each element is a matrix: genesXcells -
Output: combines data into one matrix - genesXcells}
\usage{
combine.subset(
  data_list,
  names.set = names(data_list),
  return.list = F,
  intersect = F,
  subset.genes = NULL
)
}
\description{
Combine data lists - each element is a matrix: genesXcells -
Output: combines data into one matrix - genesXcells
}
