% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Compare.GeneActivity}
\alias{Compare.GeneActivity}
\title{Identifying Unique genes sets (active/inactive/unknown)
input - list of output of genes.activity.summary for multiple data sets
output - list of unique active/inactive/unknown genes for each data set}
\usage{
\method{Compare}{GeneActivity}(data_list, all.comparisons = F)
}
\description{
Identifying Unique genes sets (active/inactive/unknown)
input - list of output of genes.activity.summary for multiple data sets
output - list of unique active/inactive/unknown genes for each data set
}
