% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{compute_pval}
\alias{compute_pval}
\title{compute p-val and q vals
input- threshold list; query gene expression matrix
output- list: q values for gene being active, q values for gene being inactive}
\usage{
compute_pval(th_list, gexp_mat, genes, p_correct = T)
}
\description{
compute p-val and q vals
input- threshold list; query gene expression matrix
output- list: q values for gene being active, q values for gene being inactive
}
