% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{createThresholdDist}
\alias{createThresholdDist}
\title{Generate threshold distributions per gene
input- vector of gene expression values
output- matrix: numbootstrap x (lower,upper thresholds)}
\usage{
createThresholdDist(gexp, numBootstrapSamples = 1000, swap = TRUE)
}
\description{
Generate threshold distributions per gene
input- vector of gene expression values
output- matrix: numbootstrap x (lower,upper thresholds)
}
