% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{cutoff.select}
\alias{cutoff.select}
\title{To select appropriate cutoff values for number of cells given gexp data
input- data: gene expression data; vector of integer values with number of cells; vector of integer values with respective cutoff values (in percentage)
output- interger - cutoff value (in percentage)}
\usage{
cutoff.select(
  data,
  ncells_vec = c(50, 250, 500, 1000),
  cutoffpct_vec = c(40, 30, 20, 10)
)
}
\description{
To select appropriate cutoff values for number of cells given gexp data
input- data: gene expression data; vector of integer values with number of cells; vector of integer values with respective cutoff values (in percentage)
output- interger - cutoff value (in percentage)
}
