% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{DAG}
\alias{DAG}
\title{Identifying differentially activated genes (DAG)
input- list of gene information (output of gene.activity.summary$geneinfo) for the two data sets that are compared}
\usage{
DAG(data_list, cutoff = 30, write.files = T, fileid = "myQuery")
}
\description{
Identifying differentially activated genes (DAG)
input- list of gene information (output of gene.activity.summary$geneinfo) for the two data sets that are compared
}
