% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{Data.gexp.stats}
\alias{Data.gexp.stats}
\title{Gene expression level computed from raw data
input- raw counts data matrix: genes x cells; threshold vector: a vector of integer values which are bins for classifying counts data; percentage of min cells required to express a given gexp value
output- named vector: names-gene name, value-maximum gene expression.
A gene is highly expressed if there exists at least min.cells percent of cell population that have above th.gexp UMI counts. Otherwise gene is lowly expressed}
\usage{
Data.gexp.stats(data_counts, th.gexp_vec, min.cells)
}
\description{
Gene expression level computed from raw data
input- raw counts data matrix: genes x cells; threshold vector: a vector of integer values which are bins for classifying counts data; percentage of min cells required to express a given gexp value
output- named vector: names-gene name, value-maximum gene expression.
A gene is highly expressed if there exists at least min.cells percent of cell population that have above th.gexp UMI counts. Otherwise gene is lowly expressed
}
