% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Data_prep1}
\alias{Data_prep1}
\title{data prep: removing genes with no expression, removing zero gexp values, scaling with maximum gene expression, removing outliers
input- normalised gexp matrix(genes x cells)
output- list with precessed data, scaling factors, outliers, genes not expressed and genes with low counts}
\usage{
Data_prep1(
  data,
  filter.genesNotExpressed = T,
  remove.zeros = T,
  scale.WithMaxGexp = T,
  remove.outliers = T
)
}
\description{
data prep: removing genes with no expression, removing zero gexp values, scaling with maximum gene expression, removing outliers
input- normalised gexp matrix(genes x cells)
output- list with precessed data, scaling factors, outliers, genes not expressed and genes with low counts
}
