% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{Data_prep2}
\alias{Data_prep2}
\title{data prep: scaling with maximum gene expression, removing outliers
input: output of background sampling - list of genes with corresponding gexp
output: list of scaled, filtered data, scaling factors, outliers}
\usage{
Data_prep2(data, scale.WithMaxGexp = T, remove.outliers = T)
}
\description{
data prep: scaling with maximum gene expression, removing outliers
input: output of background sampling - list of genes with corresponding gexp
output: list of scaled, filtered data, scaling factors, outliers
}
