% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Discretise_data}
\alias{Discretise_data}
\title{computing q values and discretising query data}
\usage{
Discretise_data(
  query_ScaledData,
  Ref_data,
  th = 0.5,
  th_i = 0.5,
  file.id = "myquery",
  return.data = T,
  saveData = T
)
}
\description{
computing q values and discretising query data
}
