% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{gene.activity.summary}
\alias{gene.activity.summary}
\title{Summary of discretised data: information for every gene, list of active, inactive, unknown genes
input- discretised data matrix: genes x cells ; cutoff (default selected by cutoff.select)
output- list: 1.geneinfo: matrix genes X gene info; 2.genelist: vectors with names of active, inactive, unknown genes}
\usage{
gene.activity.summary(
  data_dis,
  cutoff = cutoff.select(data_dis),
  return.sum = c("all", "genelist", "geneinfo")
)
}
\description{
Summary of discretised data: information for every gene, list of active, inactive, unknown genes
input- discretised data matrix: genes x cells ; cutoff (default selected by cutoff.select)
output- list: 1.geneinfo: matrix genes X gene info; 2.genelist: vectors with names of active, inactive, unknown genes
}
