% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{GetThreshDist}
\alias{GetThreshDist}
\title{Computing threshold distributions}
\usage{
GetThreshDist(
  data,
  parallelize = F,
  ncores = detectCores() - 2,
  file.id,
  return = T,
  saveData = T
)
}
\description{
Computing threshold distributions
}
