% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{LogNorm}
\alias{LogNorm}
\title{data normalisation by log transformation
input- raw data matrix - gene counts x cells, scaling factor (default=10000)
output- log normalised matrix : genes x cells}
\usage{
LogNorm(data_mat, s.f = 10000)
}
\description{
data normalisation by log transformation
input- raw data matrix - gene counts x cells, scaling factor (default=10000)
output- log normalised matrix : genes x cells
}
