% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PrepQueryData}
\alias{PrepQueryData}
\title{Normalising, Scaling query data}
\usage{
PrepQueryData(
  query_dat,
  Ref_data,
  file.id = "query",
  normalise = T,
  scale_data = T,
  saveData = T,
  return.data = T
)
}
\description{
Normalising, Scaling query data
}
