% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{PrepRefData}
\alias{PrepRefData}
\title{Sampling, scaling and removing outliers}
\usage{
PrepRefData(
  data_norm,
  metadata,
  clusterid_vec = c("Tissue", "clusterid.1", "clusterid.2"),
  file.id = "mydata",
  saveData = T,
  return = T,
  sample_data = T,
  sample_n = 50,
  data_prep = 2
)
}
\description{
Sampling, scaling and removing outliers
}
