% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read.genelist}
\alias{read.genelist}
\title{Read a file such that each column is a list of genes including NAs
Output: list with each element corresponding to the column in}
\usage{
read.genelist(file, na.rm = T)
}
\description{
Read a file such that each column is a list of genes including NAs
Output: list with each element corresponding to the column in
}
