% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{scaleMaxGexp}
\alias{scaleMaxGexp}
\title{Scaling gene expression data with given scaling factors
input- scaling factors: matrix - rownames=gene symbols, column-value of scaling factor; data matrix: genes X cells
output- scaled gene expression matrix: genes X cells}
\usage{
scaleMaxGexp(data_mat, scalingfactors, genes)
}
\description{
Scaling gene expression data with given scaling factors
input- scaling factors: matrix - rownames=gene symbols, column-value of scaling factor; data matrix: genes X cells
output- scaled gene expression matrix: genes X cells
}
