% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{scrb_norm}
\alias{scrb_norm}
\title{Data Normalisation by ScTransform}
\usage{
scrb_norm(
  data_counts,
  log_transform = T,
  data = c("Ref", "query"),
  file.id = "mydata",
  saveData = T,
  return = T
)
}
\description{
Data Normalisation by ScTransform
}
