% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Split.Data}
\alias{Split.Data}
\title{Split given data by given split parameter, return query or background data as required. Default: returns data that can be queried
input: clusterids: matrix with first column containing cell names corresponding to gexp data matrix col names}
\usage{
Split.Data(
  data_mat,
  clusterids,
  use.clusterid = "cluster.id",
  pick.clusters = NULL,
  return.onlyGxp = F,
  return.data = "Q"
)
}
\description{
Split given data by given split parameter, return query or background data as required. Default: returns data that can be queried
input: clusterids: matrix with first column containing cell names corresponding to gexp data matrix col names
}
