% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write.genelist}
\alias{write.genelist}
\title{Read a file such that each column is a list of genes including NAs
Output: list with each element corresponding to the column in}
\usage{
write.genelist(vec_list, filename, rownames = T)
}
\description{
Read a file such that each column is a list of genes including NAs
Output: list with each element corresponding to the column in
}
